split create_object_visitor.cpp to object_visitor.cpp/h
