+++ runtime/runtime.cmake
@@ -8,6 +8,7 @@ prepend(KPHP_RUNTIME_MEMORY_RESOURCE_SOURCES memory_resource/
         unsynchronized_pool_resource.cpp)
 
 prepend(KPHP_RUNTIME_MSGPACK_SOURCES msgpack/
+        object_visitor.cpp
         zone.cpp)
 
 prepend(KPHP_RUNTIME_JOB_WORKERS_SOURCES job-workers/
